+++ pkg/addons/kubectl_test.go
